Fixed MAJOR bug where after dot_life_experied the new random dot positions were position only in one quadrant of the cube. NOw uniformly. Added feature 'dot_life_sync' (default false) which controls if the dots should be refreshed all at once after 'dot_life_fr' has past or that they should update asynchronously
video size is now controlled with wid_px and hei_px instead of vid_widhei_px
Figure window size can now be controlled independently wioth fig_pos_px (eg. [100 200 300 400] for a a window with width 300 and height 400 with lower left at x=100 and y=200. If left empty, optic_flow_dots.m attempts to choose a sensible fig_pos_px based on the wid_px and hei_px paramters, and the length of the figure name (to keep that string visible)